requirements.txt
torch>=2.1.0
torchvision>=0.17.0
numpy>=1.25.0
Pillow>=10.0.0
pyyaml>=6.0.3